// src/app/models/etapa-inscricao.enum.ts
export enum EEtapaInscricao {
    AceiteIndisponivel = 0,
    AceiteDisponivel = 1,
    PendentePagamento = 2,
    ProcessandoPagamento = 3,
    PagamentoRecebido = 4,
    Matriculado = 5,
    PendenteFormaIngresso = 6,
    PendenteTaxaInscricao = 7,
    PreInscricao = 8,
    ReAceiteDisponivel = 9,
    AguardandoClassificacaoChamada = 10,
    ProcessoSeletivoEncerrado = 11,
    TaxaInscricaoInsenta = 12
}